:recycle: Move functions to base class
