+++ src/openpersonen/api/data_classes/ouder.py
@@ -1,8 +1,5 @@
 from dataclasses import dataclass
 
-from django.conf import settings
-from django.utils.module_loading import import_string
-
 from openpersonen.api.enum import GeslachtsaanduidingChoices, OuderAanduiding
 
 from .datum import Datum
@@ -17,23 +14,10 @@ class Ouder(Persoon):
     datumIngangFamilierechtelijkeBetrekking: Datum
     inOnderzoek: OuderInOnderzoek
 
+    backend_function_name = "get_ouder"
+
     def get_geslachtsaanduiding_display(self):
         return GeslachtsaanduidingChoices.values[self.geslachtsaanduiding]
 
     def get_ouderAanduiding_display(self):
         return OuderAanduiding.values[self.ouderAanduiding]
-
-    @classmethod
-    def list(cls, bsn):
-        class_instances = []
-        backend = import_string(settings.OPENPERSONEN_BACKEND)
-        instance_dicts = backend.get_ouder(bsn)
-        for instance_dict in instance_dicts:
-            class_instances.append(cls(**instance_dict))
-        return class_instances
-
-    @classmethod
-    def retrieve(cls, bsn, id):
-        backend = import_string(settings.OPENPERSONEN_BACKEND)
-        instance_dicts = backend.get_ouder(bsn, ouder_bsn=id)
-        return cls(**instance_dicts[0])
