+++ src/openpersonen/contrib/stufbg/backend.py
@@ -19,17 +19,17 @@ class BackEnd:
 
         return convert_response_to_persoon_dicts(response)
 
-    def get_kind(self, bsn, kind_bsn=None):
+    def get_kind(self, bsn, id=None):
         StufBGClient = apps.get_model("stufbg", "StufBGClient")
         response = StufBGClient.get_solo().get_kind(bsn)
-        return convert_response_to_kind_dict(response, id=kind_bsn)
+        return convert_response_to_kind_dict(response, id=id)
 
-    def get_ouder(self, bsn, ouder_bsn=None):
+    def get_ouder(self, bsn, id=None):
         StufBGClient = apps.get_model("stufbg", "StufBGClient")
         response = StufBGClient.get_solo().get_ouder(bsn)
-        return convert_response_to_ouder_dict(response, id=ouder_bsn)
+        return convert_response_to_ouder_dict(response, id=id)
 
-    def get_partner(self, bsn, partner_bsn=None):
+    def get_partner(self, bsn, id=None):
         StufBGClient = apps.get_model("stufbg", "StufBGClient")
         response = StufBGClient.get_solo().get_partner(bsn)
-        return convert_response_to_partner_dict(response, id=partner_bsn)
+        return convert_response_to_partner_dict(response, id=id)
